# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "LifeLink API",
    "version": "1.0.0",
    "description": "OpenAPI documentation for the LifeLink donation platform API. Primary routes are mounted at root paths and compatibility aliases are also available under /api/v1/*."
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Configured server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/hospital/donations": {
      "get": {
        "tags": [
          "Hospital"
        ],
        "summary": "List donations for the authenticated hospital's requests",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "pending",
                "scheduled",
                "completed",
                "cancelled"
              ]
            }
          },
          {
            "in": "query",
            "name": "skip",
            "schema": {
              "type": "integer",
              "example": 0
            },
            "description": "Legacy pagination alias still supported."
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "example": 10
            }
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Donations retrieved successfully"
          },
          "401": {
            "description": "Missing or invalid JWT"
          },
          "403": {
            "description": "Role not allowed"
          }
        }
      }
    },
...
    "/admin/hospitals": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "List hospitals",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Hospital list"
          }
        }
      }
    },
...
    "/rewards/points": {
      "get": {
        "summary": "Get donor's points summary and tier",
        "tags": [
          "Rewards - Donor"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Points summary with tier info"
          }
        }
      }
    },
...
    "/notifications": {
      "get": {
        "summary": "List notifications for authenticated user",
        "tags": [
          "Notifications"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20
            }
          },
          {
            "in": "query",
            "name": "read",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "type",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Notifications list"
          }
        }
      }
    },
...
    "/hospitals": {
      "get": {
        "summary": "List hospitals for discovery",
        "tags": [
          "Discovery"
        ],
        "description": "Public discovery endpoint. Compatibility alias is also mounted under `/api/v1/hospitals`.",
        "parameters": [
          {
            "in": "query",
            "name": "city",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "governorate",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20
            }
          },
          {
            "in": "query",
            "name": "skip",
            "schema": {
              "type": "integer"
            },
            "description": "Legacy pagination alias still supported for compatibility."
          }
        ],
        "responses": {
          "200": {
            "description": "Hospitals list"
          }
        }
      }
    },
...
    "/donations/book-appointment": {
      "post": {
        "tags": [
          "Appointments"
        ],
        "summary": "Book a donor appointment with a hospital",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "hospitalId",
                  "appointmentDate"
                ],
                "properties": {
                  "hospitalId": {
                    "type": "string"
                  },
                  "requestId": {
                    "type": "string",
                    "nullable": true
                  },
                  "appointmentDate": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "notes": {
                    "type": "string",
                    "maxLength": 500
                  }
                }
              },
              "example": {
                "hospitalId": "66f100000000000000000010",
                "requestId": "66f100000000000000000020",
                "appointmentDate": "2026-05-10T10:00:00.000Z",
                "notes": "First-time donor, available in the morning."
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Appointment booked"
          },
          "400": {
            "description": "Invalid appointment payload"
          },
          "401": {
            "description": "Missing or invalid JWT"
          },
          "403": {
            "description": "Role not allowed"
          },
          "404": {
            "description": "Hospital, donor, or request not found"
          },
          "409": {
            "description": "Active appointment already exists for this hospital"
          }
        }
      }
    },
    "/donations/book-appointment/my-appointments": {
      "get": {
        "tags": [
          "Appointments"
        ],
        "summary": "Get the authenticated donor appointments",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "skip",
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Appointments fetched successfully"
          },
          "401": {
            "description": "Missing or invalid JWT"
          },
          "403": {
            "description": "Role not allowed"
          }
        }
      }
    },
    "/donations/book-appointment/{appointmentId}": {
      "delete": {
        "tags": [
          "Appointments"
        ],
        "summary": "Cancel a donor appointment",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "appointmentId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Appointment cancelled"
          },
          "400": {
            "description": "Invalid appointment id"
          },
          "401": {
            "description": "Missing or invalid JWT"
          },
          "403": {
            "description": "Role not allowed"
          },
          "404": {
            "description": "Appointment not found"
          }
        }
      }
    },
...
  },
...
}